[FEAT] AnimeDetailClientPage 컴포넌트 추가 #
